+++ src/api/src/renderer_impl.h
@@ -26,7 +26,7 @@ class Renderer::Impl {
   GaussianSplats load_from_ply(const std::string& path);
 
   RenderedImage draw(GaussianSplats splats, const float* view, const float* projection, uint32_t width, uint32_t height,
-                     uint8_t* dst);
+                     float* dst);
 
  private:
   std::shared_ptr<core::Renderer> renderer_;
